Reduce the number of chain specs
