--- node/service/src/chain_spec/frequency_local.rs
@@ -1,154 +1,0 @@
-#![allow(missing_docs)]
-use common_primitives::node::AccountId;
-
-use common_runtime::constants::{
-	currency::EXISTENTIAL_DEPOSIT, FREQUENCY_LOCAL_TOKEN, TOKEN_DECIMALS,
-};
-
-use cumulus_primitives_core::ParaId;
-use frequency_runtime::{AuraId, CouncilConfig, Ss58Prefix, SudoConfig, TechnicalCommitteeConfig};
-use sc_service::ChainType;
-use sp_core::sr25519;
-use sp_runtime::traits::AccountIdConversion;
-
-use super::{get_account_id_from_seed, get_collator_keys_from_seed, get_properties, Extensions};
-/// Specialized `ChainSpec` for the normal parachain runtime.
-pub type ChainSpec = sc_service::GenericChainSpec<frequency_runtime::GenesisConfig, Extensions>;
-
-/// Generate the session keys from individual elements.
-///
-/// The input must be a tuple of individual keys (a single arg for now since we have just one key).
-pub fn template_session_keys(keys: AuraId) -> frequency_runtime::SessionKeys {
-	frequency_runtime::SessionKeys { aura: keys }
-}
-
-pub fn local_testnet_config() -> ChainSpec {
-	// Give your base currency a unit name and decimal places
-	let properties =
-		get_properties(FREQUENCY_LOCAL_TOKEN, TOKEN_DECIMALS as u32, Ss58Prefix::get().into());
-
-	ChainSpec::from_genesis(
-		// Name
-		"Frequency Local Testnet",
-		// ID
-		"frequency-local",
-		ChainType::Local,
-		move || {
-			testnet_genesis(
-				// initial collators.
-				vec![
-					(
-						get_account_id_from_seed::<sr25519::Public>("Alice"),
-						get_collator_keys_from_seed("Alice"),
-					),
-					(
-						get_account_id_from_seed::<sr25519::Public>("Bob"),
-						get_collator_keys_from_seed("Bob"),
-					),
-				],
-				// Sudo
-				Some(get_account_id_from_seed::<sr25519::Public>("Alice")),
-				// Endowed Accounts
-				vec![
-					get_account_id_from_seed::<sr25519::Public>("Alice"),
-					get_account_id_from_seed::<sr25519::Public>("Bob"),
-					get_account_id_from_seed::<sr25519::Public>("Charlie"),
-					get_account_id_from_seed::<sr25519::Public>("Dave"),
-					get_account_id_from_seed::<sr25519::Public>("Eve"),
-					get_account_id_from_seed::<sr25519::Public>("Ferdie"),
-					get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
-					get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
-					get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
-					get_account_id_from_seed::<sr25519::Public>("Dave//stash"),
-					get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
-					get_account_id_from_seed::<sr25519::Public>("Ferdie//stash"),
-					common_runtime::constants::TREASURY_PALLET_ID.into_account_truncating(),
-				],
-				// Council members
-				vec![
-					get_account_id_from_seed::<sr25519::Public>("Alice"),
-					get_account_id_from_seed::<sr25519::Public>("Charlie"),
-					get_account_id_from_seed::<sr25519::Public>("Eve"),
-				],
-				// Technical Committee members
-				vec![
-					get_account_id_from_seed::<sr25519::Public>("Bob"),
-					get_account_id_from_seed::<sr25519::Public>("Dave"),
-				],
-				// ParaId
-				2000.into(),
-			)
-		},
-		// Bootnodes
-		Vec::new(),
-		// Telemetry
-		None,
-		// Protocol ID
-		Some("frequency-local"),
-		// Fork ID
-		None,
-		// Properties
-		Some(properties),
-		// Extensions
-		Extensions {
-			relay_chain: "rococo-local".into(), // You MUST set this to the correct network!
-			para_id: 2000,
-		},
-	)
-}
-
-fn testnet_genesis(
-	invulnerables: Vec<(AccountId, AuraId)>,
-	root_key: Option<AccountId>,
-	endowed_accounts: Vec<AccountId>,
-	council_members: Vec<AccountId>,
-	technical_committee_members: Vec<AccountId>,
-	id: ParaId,
-) -> frequency_runtime::GenesisConfig {
-	frequency_runtime::GenesisConfig {
-		system: frequency_runtime::SystemConfig {
-			code: frequency_runtime::WASM_BINARY
-				.expect("WASM binary was not build, please build it!")
-				.to_vec(),
-		},
-		balances: frequency_runtime::BalancesConfig {
-			balances: endowed_accounts.iter().cloned().map(|k| (k, 1 << 60)).collect(),
-		},
-		parachain_info: frequency_runtime::ParachainInfoConfig { parachain_id: id },
-		collator_selection: frequency_runtime::CollatorSelectionConfig {
-			invulnerables: invulnerables.iter().cloned().map(|(acc, _)| acc).collect(),
-			candidacy_bond: EXISTENTIAL_DEPOSIT * 16,
-			..Default::default()
-		},
-		session: frequency_runtime::SessionConfig {
-			keys: invulnerables
-				.into_iter()
-				.map(|(acc, aura)| {
-					(
-						acc.clone(),                 // account id
-						acc,                         // validator id
-						template_session_keys(aura), // session keys
-					)
-				})
-				.collect(),
-		},
-		// no need to pass anything to aura, in fact it will panic if we do. Session will take care
-		// of this.
-		aura: Default::default(),
-		aura_ext: Default::default(),
-		parachain_system: Default::default(),
-		sudo: SudoConfig {
-			// Assign network admin rights.
-			key: root_key,
-		},
-		schemas: Default::default(),
-		vesting: Default::default(),
-		democracy: Default::default(),
-		treasury: Default::default(),
-		council: CouncilConfig { phantom: Default::default(), members: council_members },
-		technical_committee: TechnicalCommitteeConfig {
-			phantom: Default::default(),
-			members: technical_committee_members,
-		},
-	}
-}
