+++ node/service/src/chain_spec/frequency_rococo.rs
@@ -1,20 +1,19 @@
 #![allow(missing_docs)]
 use common_primitives::node::AccountId;
 use common_runtime::constants::{
-	currency::EXISTENTIAL_DEPOSIT, FREQUENCY_ROCOCO_TOKEN, TOKEN_DECIMALS,
+	currency::EXISTENTIAL_DEPOSIT, FREQUENCY_LOCAL_TOKEN, TOKEN_DECIMALS,
 };
 use cumulus_primitives_core::ParaId;
 use frequency_runtime::{AuraId, CouncilConfig, Ss58Prefix, SudoConfig, TechnicalCommitteeConfig};
-use hex::FromHex;
 use sc_service::ChainType;
-use sc_telemetry::TelemetryEndpoints;
-use sp_core::ByteArray;
+use sp_core::sr25519;
 use sp_runtime::traits::AccountIdConversion;
 /// Specialized `ChainSpec` for the normal parachain runtime.
 pub type ChainSpec = sc_service::GenericChainSpec<frequency_runtime::GenesisConfig, Extensions>;
 
-use super::{get_properties, Extensions};
+use super::{get_account_id_from_seed, get_collator_keys_from_seed, get_properties, Extensions};
 
+/// Generates the Live Frequency Rococo chain spec from the raw json
 pub fn load_frequency_rococo_spec() -> ChainSpec {
 	ChainSpec::from_json_bytes(
 		&include_bytes!("../../../../resources/frequency-rococo.raw.json")[..],
@@ -22,162 +21,91 @@ pub fn load_frequency_rococo_spec() -> ChainSpec {
 	.unwrap()
 }
 
-pub mod public_testnet_keys {
-	pub const COLLATOR_1_SR25519: &str =
-		"0x5c0f55ba602f76d69b5cc075d81f6d27db9157c90dc4be492f4edbe7d7c96d18";
-	pub const COLLATOR_2_SR25519: &str =
-		"0x202be6542ed50679271280b8c37140681bbd5acfd0e508668297029f871b9f0c";
-	pub const ROCOCO_FRQ_SUDO: &str =
-		"0xccca4a5b784105460c5466cbb8d11b34f29ffcf6c725d07b65940e697763763c";
-	pub const TECH_COUNCIL1: &str =
-		"0x847c1ac02474b90cf1e9d8e722318b75fd56d370e6f35e9c983fe671e788d23a";
-	pub const TECH_COUNCIL2: &str =
-		"0x52b580c22c5ff6f586a0966fbd2373de279d1aa1b2d05dff47616b5a338fce27";
-	pub const TECH_COUNCIL3: &str =
-		"0x6a13f08b279cb33b249954190bcee832747b9aa9dc14cc290f82d73d496cfc0a";
-	pub const FRQ_COUNCIL1: &str =
-		"0xa608f3e0030c157b6e2a540c5f0c7dbd6004793813cad2c9fbda0c84c093c301";
-	pub const FRQ_COUNCIL2: &str =
-		"0x52d76db441043a5d47d9bf83e6fd2d5acb86b8547062571ee7b68255b6bada10";
-	pub const FRQ_COUNCIL3: &str =
-		"0x809d0a4e6683ebff9d74c7f6ba9fe504a64a7227d74eb45ee85556cc01013a63";
-	pub const FRQ_COUNCIL4: &str =
-		"0x8e47c13fd0f028f56378e202523fa44508fd64df89fddb482fc0b128989e9f0b";
-	pub const FRQ_COUNCIL5: &str =
-		"0xf23d555b95ca8c752b531e48848bfb4d3aa2b4eea407484ccee947501e77d04f";
-	pub const FRQ_COUNCIL6: &str =
-		"0xe87a126794cb727b5a7760922f81fbf0f80fd64b7e86e6ae4fee0be4289c7512";
-	pub const FRQ_COUNCIL7: &str =
-		"0x14a6bff08e9637457a165779765417feca01a2119dec98ec134f8ae470111318";
-	pub const FRQ_COUNCIL8: &str =
-		"0x140c17ced6e4fba8b62a6935052cfb7c5a8ad8ecc43dee1f4fc7c30c1ca3cb14";
-	pub const FRQ_COUNCIL9: &str =
-		"0xfc61655783e14b361d2b9601c657c3c5361a2cf32aa1a448fc83b1a356808a1a";
-	pub const FRQ_COUNCIL10: &str =
-		"0x1caccabbe8095a35292782cffc29145030264f9706753923a5c237e8f5aceb1a";
+/// Generate the session keys from individual elements.
+///
+/// The input must be a tuple of individual keys (a single arg for now since we have just one key).
+pub fn template_session_keys(keys: AuraId) -> frequency_runtime::SessionKeys {
+	frequency_runtime::SessionKeys { aura: keys }
 }
 
-pub fn frequency_rococo_testnet() -> ChainSpec {
+/// Generates the chain spec for a local testnet
+pub fn local_testnet_config() -> ChainSpec {
+	// Give your base currency a unit name and decimal places
 	let properties =
-		get_properties(FREQUENCY_ROCOCO_TOKEN, TOKEN_DECIMALS as u32, Ss58Prefix::get().into());
-	let para_id: ParaId = 4044.into();
+		get_properties(FREQUENCY_LOCAL_TOKEN, TOKEN_DECIMALS as u32, Ss58Prefix::get().into());
+
 	ChainSpec::from_genesis(
 		// Name
-		"Frequency Rococo",
+		"Frequency Local Testnet",
 		// ID
-		"frequency-rococo",
-		ChainType::Live,
+		"frequency-local",
+		ChainType::Local,
 		move || {
-			frequency_rococo_genesis(
+			testnet_genesis(
 				// initial collators.
 				vec![
 					(
-						// 5FLZcmvR22FvM2u4GDbeSAQ8uJpePuaEQccquD89UcVBwEgB
-						public_testnet_keys::COLLATOR_1_SR25519
-							.parse::<AccountId>()
-							.unwrap()
-							.into(),
-						AuraId::from_slice(
-							&<[u8; 32]>::from_hex(
-								public_testnet_keys::COLLATOR_1_SR25519.strip_prefix("0x").unwrap(),
-							)
-							.unwrap(),
-						)
-						.unwrap(),
+						get_account_id_from_seed::<sr25519::Public>("Alice"),
+						get_collator_keys_from_seed("Alice"),
 					),
 					(
-						// 5CkeZ1PLWpdS4aSYKBLxVwzZGfvGa2SitSKhX72KT1KAih4Z
-						public_testnet_keys::COLLATOR_2_SR25519
-							.parse::<AccountId>()
-							.unwrap()
-							.into(),
-						AuraId::from_slice(
-							&<[u8; 32]>::from_hex(
-								public_testnet_keys::COLLATOR_2_SR25519.strip_prefix("0x").unwrap(),
-							)
-							.unwrap(),
-						)
-						.unwrap(),
+						get_account_id_from_seed::<sr25519::Public>("Bob"),
+						get_collator_keys_from_seed("Bob"),
 					),
 				],
-				Some(public_testnet_keys::ROCOCO_FRQ_SUDO.parse::<AccountId>().unwrap().into()),
-				// endowed accounts.
+				// Sudo
+				Some(get_account_id_from_seed::<sr25519::Public>("Alice")),
+				// Endowed Accounts
 				vec![
-					// 5GhDid9V5rpReBSAK6sSEmCzeiUXTgAmAoMJQUuK5gFzPXoq
-					public_testnet_keys::ROCOCO_FRQ_SUDO.parse::<AccountId>().unwrap().into(),
-					// 5E46myWYmywo8cF9Wk77NZM7TqLqVG7uMYjGuyyfrve9waa9
-					public_testnet_keys::TECH_COUNCIL1.parse::<AccountId>().unwrap().into(),
-					// 5Dw9hdb8Wy3cMFf1ZpQh5HuK2GvMNMcxfka4JsM2GRpjqHSA
-					public_testnet_keys::TECH_COUNCIL2.parse::<AccountId>().unwrap().into(),
-					// 5ETnrcfDwCTKZt5pLoKcLNbbwUcddkDAxEjtcgibAtkDhBBe
-					public_testnet_keys::TECH_COUNCIL3.parse::<AccountId>().unwrap().into(),
-					// 5FpQTxEFHAdF2hpkA89qKenSjGQBAHi8uCo8F6fJ5RxnR1sn
-					public_testnet_keys::FRQ_COUNCIL1.parse::<AccountId>().unwrap().into(),
-					// 5DwKn9yg7Kqa71qTaZVfoigwtzyPgBp23wzJKmGiUr8S3j7A
-					public_testnet_keys::FRQ_COUNCIL2.parse::<AccountId>().unwrap().into(),
-					// 5EyLduUEpNfUox5HuogrhbKugAdHnsLHrysKSWL6u4TQFAra
-					public_testnet_keys::FRQ_COUNCIL3.parse::<AccountId>().unwrap().into(),
-					// 5FHFyMYQH39mq3FA2a5ZcDe275K7RZs3zi76gsQb2TE1PF7i
-					public_testnet_keys::FRQ_COUNCIL4.parse::<AccountId>().unwrap().into(),
-					// 5HYKfXdGi2GUupHksmQLtHoWtwH3wPkJRSr3RpbKivTovpqX
-					public_testnet_keys::FRQ_COUNCIL5.parse::<AccountId>().unwrap().into(),
-					// 5HKXEFkNLeutZ2yZ9mbG2v5AbTDMXhza9VEfeDCiJGGBUyi3
-					public_testnet_keys::FRQ_COUNCIL6.parse::<AccountId>().unwrap().into(),
-					// 5CXnMFFgruHxAFtvCogwm8TFygWg1MWd9KhMnfEPbRdCpf74
-					public_testnet_keys::FRQ_COUNCIL7.parse::<AccountId>().unwrap().into(),
-					// 5CWzQaAJFqYoF1bZWsvEnzMQDGokk2csTFBwfPpo1oNfBGkn
-					public_testnet_keys::FRQ_COUNCIL8.parse::<AccountId>().unwrap().into(),
-					// 5HmcreGLq25iA7fiyb6An4YVWC3k1Cq8SQgYn2Qpeepq24nV
-					public_testnet_keys::FRQ_COUNCIL9.parse::<AccountId>().unwrap().into(),
-					// 5CiJXZxsko8YZGhTZAwrhREUzjCVEGwNE7Fgrf1twbt8WsmY
-					public_testnet_keys::FRQ_COUNCIL10.parse::<AccountId>().unwrap().into(),
-					// Treasury $$
+					get_account_id_from_seed::<sr25519::Public>("Alice"),
+					get_account_id_from_seed::<sr25519::Public>("Bob"),
+					get_account_id_from_seed::<sr25519::Public>("Charlie"),
+					get_account_id_from_seed::<sr25519::Public>("Dave"),
+					get_account_id_from_seed::<sr25519::Public>("Eve"),
+					get_account_id_from_seed::<sr25519::Public>("Ferdie"),
+					get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Dave//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Ferdie//stash"),
 					common_runtime::constants::TREASURY_PALLET_ID.into_account_truncating(),
 				],
-				// Council Members
+				// Council members
 				vec![
-					public_testnet_keys::FRQ_COUNCIL1.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL2.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL3.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL4.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL5.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL6.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL7.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL8.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL9.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::FRQ_COUNCIL10.parse::<AccountId>().unwrap().into(),
+					get_account_id_from_seed::<sr25519::Public>("Alice"),
+					get_account_id_from_seed::<sr25519::Public>("Charlie"),
+					get_account_id_from_seed::<sr25519::Public>("Eve"),
 				],
-				// Technical Committee Members
+				// Technical Committee members
 				vec![
-					public_testnet_keys::TECH_COUNCIL1.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::TECH_COUNCIL2.parse::<AccountId>().unwrap().into(),
-					public_testnet_keys::TECH_COUNCIL3.parse::<AccountId>().unwrap().into(),
+					get_account_id_from_seed::<sr25519::Public>("Bob"),
+					get_account_id_from_seed::<sr25519::Public>("Dave"),
 				],
-				para_id,
+				// ParaId
+				2000.into(),
 			)
 		},
 		// Bootnodes
-		vec![
-			"/dns4/0.boot.rococo.frequency.xyz/tcp/30333/ws/p2p/12D3KooWArmKDbY8Y6XXHGodosWAjRWWxSw5YxWEjSZTBNjJXVSC".parse().unwrap(),
-		],
+		Vec::new(),
 		// Telemetry
-		TelemetryEndpoints::new(vec![("wss://telemetry.frequency.xyz/submit/".into(), 0)]).ok(),
+		None,
 		// Protocol ID
-		Some("frequency-rococo"),
+		Some("frequency-local"),
 		// Fork ID
 		None,
 		// Properties
 		Some(properties),
 		// Extensions
 		Extensions {
-			relay_chain: "rococo".into(), // You MUST set this to the correct network!
-			para_id: para_id.into(),
+			relay_chain: "rococo-local".into(), // You MUST set this to the correct network!
+			para_id: 2000,
 		},
 	)
 }
 
-fn frequency_rococo_genesis(
-	initial_authorities: Vec<(AccountId, AuraId)>,
+fn testnet_genesis(
+	invulnerables: Vec<(AccountId, AuraId)>,
 	root_key: Option<AccountId>,
 	endowed_accounts: Vec<AccountId>,
 	council_members: Vec<AccountId>,
@@ -195,18 +123,18 @@ fn frequency_rococo_genesis(
 		},
 		parachain_info: frequency_runtime::ParachainInfoConfig { parachain_id: id },
 		collator_selection: frequency_runtime::CollatorSelectionConfig {
-			invulnerables: initial_authorities.iter().cloned().map(|(acc, _)| acc).collect(),
+			invulnerables: invulnerables.iter().cloned().map(|(acc, _)| acc).collect(),
 			candidacy_bond: EXISTENTIAL_DEPOSIT * 16,
 			..Default::default()
 		},
 		session: frequency_runtime::SessionConfig {
-			keys: initial_authorities
+			keys: invulnerables
 				.into_iter()
 				.map(|(acc, aura)| {
 					(
-						acc.clone(),                             // account id
-						acc,                                     // validator id
-						frequency_runtime::SessionKeys { aura }, // session keys
+						acc.clone(),                 // account id
+						acc,                         // validator id
+						template_session_keys(aura), // session keys
 					)
 				})
 				.collect(),
